Node can use its DOM Element to define itself
This is a huge refactor, to make possible what was previously impossible
by design. The decision was taken because building a paginated widget is
not possible without access to the on-screen geometry of the thing. With
the previous notion of VirtualElement that just could not be had: a
VirtualElement can be drawn in many places, i.e. there's no single DOM
element corresponding to one VirtualElement.

This refactor also gives what I believe is a more composable interface,
but handicapped slightly because it's founded on indexed monads rather
than monads, so we don't get to use do notation unless we rebind it and
screw up all of the other monadic computation. Check out
Examples/Main.hs for a preview of widget and flow definitions.