Yet another massive redesign
As I continue to work with this project I learn more and more about what
a principled and expressive design requires. This commit brings in these
important changes:

  1. Children containers have kind (* -> *) -> *. They use some functor
     as a parameter but are not functors themselves, and therefore not
     foldable nor traversable. This is however the way forward: it
     allows us to define containers which fix type information about
     their children. Example:
       newtype Single (t :: *) (f :: * -> *) = Single (f t)
     When we choose Single (Event ()) we assert that there is precisely
     one child and it carries an Event (). That's quite useful when
     defining a Widget.

  2. Widgets are profunctors; a UI is a Widget with unit input along
     with an HTML tag. This allows greater composition of Widgets from
     other Widgets, including "knot tying" in which multiple Widgets are
     fused and their outputs are jointly used to define their inputs.

The library remains shocking complicated and is almost certainly
inaccessible to the faint of heart. However I believe this commit brings
it to a good foundation upon which more accessible interfaces can be
built.