HTML tag is indicated in type not in term
This tag is present as a parameter on ElementBuilder and Widget.
We define an OpenWidget to be one for which the tag is not constrained.
In the future, this will imply that no events have been bound on the
widget. OpenWidgets can be composed horizontally, their children
concatenated. Any non-open Widget can be made open by creating a new
Widget, with free tag parameter, containing that other Widget as its
sole static child.

Once a tag has been constrained, the programmer can use the Modification
type and Profunctor interfaces, so it remains rather versatile and
adaptable. As always, the programmer must move the input parameter to ()
in order to construct a UI, which can be rendered or used as a child.