widgetSum and supporting refactor
The challenge: define a type-safe sum combinator, in which one of two
OpenWidget's is chosen based upon the type of the input.

What should be the children container for this? Well, it depends upon
the input! If we give a Left, the container and its changes should be
isomorphic to that of the left OpenWidget, and similarly for the Right.

In order to make this happen, the intrinsic part of a Widget has been
slightly altered, and so too have the kinds of children containers.
A children container is now parameterized by the input and output types
of the intrinsic part. This is exploited in the definition of widgetSum,
in which a pattern match on the input reveals structure of the view
children.